chore: Only check for key existence when filling
In the 'add' strategy, the filltable lambda required access to all
secrets and parameters in the existing config object when updating.
Since we're only interested in existence, this commit updates the code
to only get the bare config object, not any references. This way, the
custom resource does not require access to secrets or params.